skin-lesion-mvp/ml-service/requirements.txt
# ml-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9

torch==2.7.0
torchvision==0.22.0

pillow==11.1.0
numpy==2.1.3
pydantic==2.8.2
